requirements.txt
# SkillVerse - Python Flask Dependencies
# Install using: pip install -r requirements.txt

# Core Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Real-time Communication
Flask-SocketIO==5.5.1
python-socketio>=5.12.0

# Database
SQLAlchemy>=2.0.35
psycopg2-binary>=2.9.9

# Security & Authentication
Werkzeug==3.0.1
bcrypt==4.1.2
PyJWT>=2.8.0
itsdangerous==2.1.2 

# Email validation
email-validator>=2.1.1

# Email sending
Flask-Mail==0.9.1

# Environment variables
python-dotenv==1.0.0

# Date/Time handling
python-dateutil==2.8.2
pytz==2024.1 

# OAuth & API
Authlib==1.3.0
requests>=2.32.0
urllib3>=2.0.7
certifi>=2024.2.2
Flask-Moment==1.0.5
Flask-Compress==1.14

# Production
gunicorn==21.2.0
gevent>=24.2.1

# AI Chatbot (AskVera)
groq>=0.12.0

# Assets & Media
matplotlib
Pillow
qrcode[pil]
numpy
reportlab


